# Conflicts:
#	.gitignore
#	main.cpp

+++ src/units.h
@@ -9,8 +9,8 @@ namespace units
         uint32_t pixels_per_second;
 
         Speed(uint32_t pps) : pixels_per_second{pps} { ; }
-        [[nodiscard]] static Speed MetresPerSecond(double metres) { return Speed(metres * PIXELS_PER_METRE); };
-        [[nodiscard]] static Speed KmPerHour(double km) { return Speed(km * 1000 * PIXELS_PER_METRE / 3600); }
+        [[nodiscard]] static Speed MetresPerSecond(double metres) { return Speed(static_cast<uint32_t>(metres * PIXELS_PER_METRE)); };
+        [[nodiscard]] static Speed KmPerHour(double km) { return Speed(static_cast<uint32_t>(km * 1000 * PIXELS_PER_METRE / 3600)); }
     };
     struct Time
     {
@@ -22,7 +22,7 @@ namespace units
         uint32_t pixels;
 
         constexpr Distance(uint32_t p) : pixels{p} { ; }
-        [[nodiscard]] static constexpr Distance Metres(double metres) { return Distance(metres * PIXELS_PER_METRE); }
+        [[nodiscard]] static constexpr Distance Metres(double metres) { return Distance(static_cast<uint32_t>(metres * PIXELS_PER_METRE)); }
         [[nodiscard]] operator uint32_t() const { return pixels; }
         [[nodiscard]] double meters() const { return pixels / PIXELS_PER_METRE; }
     };
